# back/services/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "陪玩平台 Gateway 服务 API 文档",
        "title": "陪玩平台 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
...
        "/api/user/companion/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的陪玩资料信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取陪玩资料",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.GetCompanionProfileResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新当前登录用户的陪玩资料信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新陪玩资料",
                "parameters": [
                    {
                        "description": "更新陪玩资料请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateCompanionProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.UpdateCompanionProfileResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
        "/api/user/companion/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新陪玩的在线状态：0=离线, 1=在线, 2=忙碌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新陪玩状态",
                "parameters": [
                    {
                        "description": "更新陪玩状态请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateCompanionStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.UpdateCompanionStatusResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
        "/api/user/login-by-code": {
            "post": {
                "description": "使用手机号和验证码登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "验证码登录",
                "parameters": [
                    {
                        "description": "验证码登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.LoginByCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.LoginByCodeResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
        "/api/user/recharge": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据订单号查询充值订单状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查询充值订单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "充值订单号",
                        "name": "orderNo",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.RechargeQueryResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建帅币充值订单，支持支付宝支付",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "创建充值订单",
                "parameters": [
                    {
                        "description": "创建充值订单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RechargeCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/types.RechargeCreateResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/types.BaseResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.BaseResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.CompanionInfo": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "description": "头像URL",
                    "type": "string"
                },
                "bio": {
                    "description": "个人简介",
                    "type": "string"
                },
                "gameSkill": {
                    "description": "游戏技能（单个游戏名称）",
                    "type": "string"
                },
                "isVerified": {
                    "description": "是否认证",
                    "type": "boolean"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "pricePerHour": {
                    "description": "每小时价格（帅币）",
                    "type": "integer"
                },
                "rating": {
                    "description": "评分（0-5分）",
                    "type": "number"
                },
                "status": {
                    "description": "状态：0=离线, 1=在线, 2=忙碌",
                    "type": "integer"
                },
                "totalOrders": {
                    "description": "总接单数",
                    "type": "integer"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "integer"
                }
            }
        },
...
        "types.GetCompanionProfileResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.CompanionInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.LoginByCodeRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "types.LoginByCodeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.LoginData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "types.LoginData": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "description": "Access Token（短期，15分钟）",
                    "type": "string"
                },
                "expiresIn": {
                    "description": "Access Token 过期时间（秒）",
                    "type": "integer"
                },
                "refreshToken": {
                    "description": "Refresh Token（长期，7天）",
                    "type": "string"
                }
            }
        },
...
        "types.RechargeCreateData": {
            "type": "object",
            "properties": {
                "expiresIn": {
                    "description": "订单有效期（秒）",
                    "type": "integer"
                },
                "orderNo": {
                    "description": "充值单号",
                    "type": "string"
                },
                "payForm": {
                    "description": "支付表单（APP/PC场景可选）",
                    "type": "string"
                },
                "payUrl": {
                    "description": "支付跳转URL（WAP/PC）",
                    "type": "string"
                }
            }
        },
        "types.RechargeCreateRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "充值金额（分/帅币）",
                    "type": "integer"
                },
                "payType": {
                    "description": "支付方式：alipay_wap / alipay_page / alipay_app",
                    "type": "string"
                },
                "returnUrl": {
                    "description": "同步回跳地址（可选）",
                    "type": "string"
                }
            }
        },
        "types.RechargeCreateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.RechargeCreateData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.RechargeQueryData": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "createdAt": {
                    "description": "创建时间（格式化时间）",
                    "type": "string"
                },
                "orderNo": {
                    "type": "string"
                },
                "paidAt": {
                    "description": "支付完成时间（格式化时间）",
                    "type": "string"
                },
                "status": {
                    "description": "订单状态：0=待支付,1=成功,2=失败,3=关闭",
                    "type": "integer"
                }
            }
        },
        "types.RechargeQueryResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.RechargeQueryData"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.UpdateCompanionProfileRequest": {
            "type": "object",
            "properties": {
                "gameSkill": {
                    "description": "游戏技能（单个游戏名称）",
                    "type": "string"
                },
                "pricePerHour": {
                    "description": "每小时价格（帅币）",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：0=离线, 1=在线, 2=忙碌",
                    "type": "integer"
                }
            }
        },
        "types.UpdateCompanionProfileResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.CompanionInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "types.UpdateCompanionStatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "状态：0=离线, 1=在线, 2=忙碌",
                    "type": "integer"
                }
            }
        },
        "types.UpdateCompanionStatusResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.CompanionInfo"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Token，格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}